requirements.txt
aiogram==3.6.0
asyncpg==0.29.0
gino==1.0.1
python-dotenv==1.0.1
SQLAlchemy==1.4.39
SQLAlchemy==1.3.24
